% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/flagship-package.r
\docType{data}
\name{flagship}
\alias{flagship}
\alias{flagship}
\alias{flagship-package}
\title{Pricing Data for Flagship Universities}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500 rows and 8 columns.}
\source{
\url{https://trends.collegeboard.org/college-pricing}
}
\usage{
flagship
}
\description{
Tuition and enrollment data from the College Board for
state flagship universities in US, 2007-2016

flagship.
}
\keyword{datasets}
